+++ CHANGELOG.md
@@ -2314,3 +2314,4 @@
 - [2026-06-25 16:45] feat(navigation): add bi-directional event bus handler for navigation
 - [2026-06-25 16:57] docs(test): update readme and API contracts for test
 - [2026-06-25 17:02] feat(theme): implement shimmer loading skeleton in theme
+- [2026-06-25 17:19] test(ui): implement widget integration tests for ui